+++ src/types.ts
@@ -53,7 +53,7 @@ export type Validator = ValidatorFunction | RegExp | ValidatorOptions | Validato
  * Defer an reference with an function (or as other projects call it "Forward declaration")
  * @param type This is just to comply with the common pattern of `type => ActualType`
  */
-export type DeferredFunc<T = any> = (type: undefined) => T;
+export type DeferredFunc<T = any> = (...args: unknown[]) => T;
 
 export interface BasePropOptions {
   [key: string]: any;